Refactor error tracking
