+++ src/App/utils/trackErrors.js
@@ -1,0 +1,22 @@
+import Honeybadger from 'honeybadger-js'
+import {honeybadgerKey} from './keys'
+
+export default (store) => {
+  
+  Honeybadger.configure({
+    api_key: honeybadgerKey,
+    onerror: false,
+  })
+
+  window.onerror = (message, url, line, column, error) => {
+    const instructorId = store.getState().appScreen.user.instructor_id
+    Honeybadger.setContext({
+      instructorId,
+    })
+    Honeybadger.notify(error ? error : {
+      name: message,
+      message,
+      stack: [message, '\n    at ? (', url || 'unknown', ':', line || 0, ':', column || 0, ')'].join(''),
+    })
+  }
+}
